Combine `generate_impl` and `generate_trait_impl` into a single file
